Bump middleware to latest suggested implementation
